% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPOWER.r
\name{DPOWER}
\alias{DPOWER}
\title{Find the week number of year}
\usage{
DPOWER(month, maxdate, curdate, bpem)
}
\arguments{
\item{month, maxdate, curdate, bpem/}{month,maxdate,curdate and bpem need to be entered}
}
\value{
return the week number in year
}
\description{
Takes in any date and show the week number
}
